Utils :  - Added boost_thread.h header to remove warnings with GCC.  - Improved Frost types compilation performance by splitting template declaration and implementation.  - Removed some classes from frost_utils.h : these will have to be included manually when needed : Color, Vector, Timer, PeriodicTimer, TimeManager, Event, EventReceiver, EventManager, AxisAlignedBox, Point, Profiler.  - Fixed some ambiguous function definitions with s_str.  - Added a flexible error management for Lua : you can now customize the error string at will using LuaManager::SetErrorFunction(). This change required a little adjustment to Lua 5.1 (in ldebug.c).
Engine :
 - Removed many headers from frost_stableheaders.h to improve compilation time. These headers are now only included when needed, thus also reducing duplicated useless code.

+++ Input/src/frost_inputmanager.cpp
@@ -7,6 +7,11 @@
 #include "frost_inputmanager.h"
 #include "frost_localemanager.h"
 
+#include <frost_utils_eventreceiver.h>
+#include <frost_utils_event.h>
+#include <frost_utils_eventmanager.h>
+#include <frost_utils_timemanager.h>
+
 #include <OIS.h>
 
 using namespace std;
@@ -498,8 +503,8 @@ namespace Frost
             bLastDragged_ = false;
 
         // Update mouse position
-        fRawDMX_ = fDMX_ = mMouseState.X.abs - fMX_;
-        fRawDMY_ = fDMY_ = mMouseState.Y.abs - fMY_;
+        fRawDMX_ = fDMX_ = (float)mMouseState.X.abs - fMX_;
+        fRawDMY_ = fDMY_ = (float)mMouseState.Y.abs - fMY_;
         /*fRawDMX_ = fDMX_ = mMouseState.X.rel;
         fRawDMY_ = fDMY_ = mMouseState.Y.rel;*/
